app/api/pyproject.toml
[tool.poetry]
name = "goat_api"
version = "2.0.0"
description = "Core API for GOAT service spatial data and different indicators."
authors = ["EPajares <elias.pajares@plan4better.de>", "Majk Shkurti <majk.shkurti@plan4better.de>", "Mohammad Etemaddar <mohammad.etemaddar@plan4better.de>"]
readme = "README.md"
keywords = ["GOAT", "API", "Spatial", "Data", "Indicators"]


[tool.black]
line-length = 88
target-version = [ "py310", "py311" ]
exclude = "((.eggs | .git | .pytest_cache | build | dist))"

[tool.ruff]
line-length = 88
exclude = [".git", "__pycache__", ".mypy_cache", ".pytest_cache"]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    # "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "B904",
    "B006",    
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.poetry.dependencies]
python = ">3.9,<3.12"
uvicorn = "^0.17.6"
fastapi = {extras = ["all"], version = "^0.96.1"}
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
celery = {extras = ["librabbitmq", "redis", "sqs"], version = "^5.2.7"}
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
pydantic = "^1.10.7"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^3.1.2"
psycopg2-binary = "^2.8.5"
alembic = "^1.4.2"
SQLAlchemy = "^1.4.23"
httpx = "^0.23.0"
asyncpg = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
GeoAlchemy2 = "^0.9.4"
geojson = "^2.5.0"
geopandas = "0.13.2"
morecantile = "^2.1.4"
cppimport = "^21.3.7"
pandas = "^2.0.2"
pytest = "^7.1.1"
sentry-sdk = "^1.5.0"
alembic_utils = "^0.7.4"
rich = "^11.0.0"
sqlmodel = "^0.0.8"
schedule = "^1.1.0"
geobuf = "^1.1.1"
schema = "^0.7.5"
openpyxl = "^3.0.9"
XlsxWriter = "^3.0.3"
PyYAML = "^6.0"
motor = "^3.0.0"
scipy = "1.11.1"
cython = "^0.29.35"
pyarrow = "^12.0.1"
asgiref = "^3.7.2"
aiohttp = "^3.8.4"
boto3 = "^1.26.164"
h3 = "^3.7.6"
numba = "^0.57.1"


[tool.poetry.group.dev.dependencies]
sqlalchemy-stubs = "^0.3"
debugpy = "^1.4.1"
ruff = "^0.0.269"
black = "^23.3.0"
pytest = "^7.3.2"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
pytest-postgresql = "^5.0.0"
pre-commit = "^3.3.3"
pytest-testmon = "^2.0.9"
pytest-sugar = "^0.9.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportGeneralTypeIssues = false

[tool.pytest.ini_options]
addopts="-n 1"